Include the file extension for `sourceFileName`.
Source maps expect `sources[]` entries to refer to the original file.
`sourceFileName` should include the file extension instead of using just
the RequireJS module name.

Including the extension also avoids conflicts between directory and file
names in common cases such as:

  - /path/to
    - module.js
    - module
      - submodule.js

Without the extension on `module.js`, the debugger will likely hide either
the file or all of the files in the directory.