Add support for the current Babel 7
This change requires editing the `paths` property of the RequireJS
configuration and edding `babel` modules as a dependency to the client
project. As such it is a breaking change and should be released as a new
major version. Althou the function `resolveModuleSource` is not necessary
to configure any more, it is supported for compatibility.

* Depend on `babel-plugin-module-resolver-standalone` as a replacement for
  the removed support for `resolveModuleSource` in Babel.
* Remove the copy of `babel-standalone`. The newest version supported by this
  plugin should be loaded by the client.
* Declare `@babel/standalone` and `babel-plugin-module-resolver-standalone`
  as dependencies to give a hint, that they have to be added by the client.
* Add another demo project demonstrating integration of `@babel/polyfill` and
  `babel-external-helpers` to be able to use the `async`/`await` syntax.
* Include package lock to follw NPM best practices.
* Do not mention bower, which has been deprecated in favour of NPM.