Update demo to test loading lit-html works correctly, and that relative paths within the original example work correctly.
You probably don't want to commit this to the repository,
or maybe want to commit it as another demo/test.

Also, note that I couldn't support a path for lit-html of

'lit-html': '../node_modules/lit-html/lit-html'

because then there's no way to specify the path to lit-html's support
files (i.e. the other files in the lit-html/ directory).  So instead
I did:

'lit-html': '../node_modules/lit-html'